Re-work platforms_test to test sensor attachment. Re-work sensor_testo to test bistatic acoustic background tasks. Create transmit_test and fix package based on these tests.
