+++ config/dictionary.txt
@@ -73,3 +73,4 @@ multistatic
 harmonic
 guassians
 refraction
+basebanded
