+++ transmit/transmit_cw.cc
@@ -22,9 +22,8 @@ transmit_cw::transmit_cw(const std::string& type, double duration,
                          double fcenter, double delay, double source_level,
                          window::type window, double window_param,
                          double orderedAZ, double orderedDE, int transmit_mode)
-    : transmit_model(type, duration, fcenter, 1.0 / duration, delay,
-                     source_level, window, window_param, orderedAZ, orderedDE,
-                     transmit_mode) {}
+    : transmit_model(type, duration, 0.0, fcenter, delay, source_level, window,
+                     window_param, orderedAZ, orderedDE, transmit_mode) {}
 
 /**
  * Creates a complex analytic signal for this waveform.
