+++ transmit/transmit_model.h
@@ -90,8 +90,8 @@ class USML_DECLSPEC transmit_model {
      *
      * @param type          Display name for waveform type.
      * @param duration      Duration of the transmitted signal (sec).
-     * @param fcenter       Center frequency for current modulation scheme (Hz).
      * @param bandwidth     Frequency bandwidth for current modulation (Hz).
+     * @param fcenter       Center frequency for current modulation scheme (Hz).
      * @param delay         Delay after completion of previous pulse (sec).
      * @param source_level  Peak intensity of pulse (microPa at 1m).
      * @param window_type   Waveform shading window.
@@ -100,8 +100,8 @@ class USML_DECLSPEC transmit_model {
      * @param orderedDE  	Ordered steering elevation relative to host (deg).
      * @param transmit_mode Transmitter beam pattern number to use.
      */
-    transmit_model(const std::string& type, double duration, double fcenter,
-                   double bandwidth, double delay, double source_level,
+    transmit_model(const std::string& type, double duration, double bandwidth, double fcenter,
+                   double delay, double source_level,
                    window::type window_type, double window_param,
                    double orderedAZ, double orderedDE, int transmit_mode);
